\name{Illumina Genomic data}
\docType{data}
\alias{chr17.260}
\alias{QC.260}
\title{Illumina example data}
\description{
These datasets are subsets of an experiment to test the applicability of 
paraffin embedded material in Illumina SNP arrays
}
\usage{
data(chr17.260)
data(QC.260)
}
\format{
\code{chr17.260} is a \code{SnpSetIllumina} object with data from chromosome 17
  of 24 samples.
\code{QC.260} is a \code{QCIllumina} object with summary data of 96 samples of a
  single SAM array

}
\keyword{datasets}
